added logger to combat

+++ roguelike/src/main/kotlin/ru/spbau/shavkunov/roguelike/attributes/Attributes.kt
@@ -59,8 +59,8 @@ data class Attributes(
         }
     }
 
-    fun absorbingDamage(damage: Int): Double {
-        return (damage * armor).toFloat() / 100.0
+    fun absorbingDamage(damage: Int): Int {
+        return damage * armor / 100
     }
 }
